+++ app/src/main/java/com/codesoom/assignment/web/controller/CommonExceptionAdvisor.java
@@ -1,7 +1,5 @@
 package com.codesoom.assignment.web.controller;
 
-import com.codesoom.assignment.web.exception.MemberNotFoundException;
-import com.codesoom.assignment.web.exception.ProductNotFoundException;
 import com.codesoom.assignment.web.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.converter.HttpMessageNotReadableException;
@@ -14,25 +12,7 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ControllerAdvice
-public class ExceptionAdvisor {
-    @ResponseBody
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(ProductNotFoundException.class)
-    public ErrorResponse handleProductNotFound() {
-        return ErrorResponse.builder()
-                .message("Product not found")
-                .build();
-    }
-
-    @ResponseBody
-    @ResponseStatus(HttpStatus.NOT_FOUND)
-    @ExceptionHandler(MemberNotFoundException.class)
-    public ErrorResponse handleMemberNotFound() {
-        return ErrorResponse.builder()
-                .message("Member not found")
-                .build();
-    }
-
+public class CommonExceptionAdvisor {
     @ResponseBody
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     @ExceptionHandler(HttpMessageNotReadableException.class)
